fix: safely exit conversion rate indexing on error
When conversion rate indexes are up to date, the indexer will attempt to index from "tomorrow". As a result, fetching throws an error. But the interval is not cleared, so the loop continues indefinitely.

This solution prevents fetching from even happening by checking the validity before fetching. It also ensures that we safely exit and end all intervals whenever an error occurs.